chore(CI): improve paths transformation
