+++ .github/workflows/ci.yml
@@ -30,7 +30,7 @@ jobs:
       - name: Run test suite
         run: |
           npm run test:coverage
-          node scripts/absolute2relative.js
+          node scripts/convertCoveragePaths.js .coverage/lcov.info
         env:
           CI: true
       - name: Save code coverage
@@ -66,7 +66,10 @@ jobs:
           name: macOS-latest.lcov
           path: .coverage/macOS
       - name: Run Sonar Scanner
-        run: sh scripts/sonar-scanner.sh
+        run: |
+          npm run build
+          node scripts/convertCoveragePaths.js .coverage/macOs/lcov.info .coverage/windows/lcov.info --back
+          sh scripts/sonar-scanner.sh
         env:
           SONAR_SCANNER_VERSION: 4.0.0.1744
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
