+++ sonar-project.properties
@@ -1,4 +1,4 @@
 sonar.organization=corpusculejs
 sonar.projectKey=corpusculejs_custom-builtin-elements
-sonar.sources=src
+sonar.sources=lib
 sonar.host.url=https://sonarcloud.io
